backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
marshmallow
marshmallow-sqlalchemy
flask-jwt-extended
python-dotenv
Flask>=3.0
flask-cors>=4.0
Flask-SQLAlchemy>=3.1
psycopg2-binary>=2.9
python-dotenv>=1.0
black
ruff
pytest
flask_bcrypt
requests-oauthlib